requirements.txt
numpy==1.24.3
opencv-python==4.8.0.76
pytesseract==0.3.10
Pillow==10.0.1
streamlit
ultralytics
torch 
torchvision